move from simply.py to __init__.py bump version, remove digicam
